* Show "You're victorious' message on victory and ask confirmation upon deleting a saved game. * Added PopUp shortcuts for ready messages.

+++ trunk/MobileClient/www/Javascript/UI/MenuScreen/SkirmishScreen.js
@@ -11,8 +11,12 @@ var SkirmishScreen = new function () {
 		MenuScreenState.selectedSaveName = $(event.currentTarget).attr('GameSlotName');
 
 		if ($(event.target).hasClass('game_slot_entry_delete_button')) {
-			SavesStorage.deleteGame(MenuScreenState.selectedSaveName);
-			reloadGameSlotsInfo();
+
+			PopUpManager.confirmYesNo('Are you sure you want to delete this saved game?', function () {
+				SavesStorage.deleteGame(MenuScreenState.selectedSaveName);
+				reloadGameSlotsInfo();
+			});
+
 			return;
 		}
